split UniversalRender into client and server render #669
