+++ app/utils/StateFunctions.js
@@ -4,7 +4,7 @@ import { parsePayoutAmount, repLog10 } from 'app/utils/ParsersAndFormatters';
 import { Long } from 'bytebuffer';
 import { VEST_TICKER, LIQUID_TICKER } from 'app/client_config';
 import { Map, Seq, fromJS } from 'immutable';
-import { getStoreState } from 'shared/UniversalRender';
+import { getStoreState } from 'app/clientRender';
 
 export const numberWithCommas = x => x.replace(/\B(?=(\d{3})+(?!\d))/g, ',')
